chore: move Reader into separate reader.py file
